move alerts into components/alerts
